api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-consul==1.1.0
confluent-kafka==2.2.0
kafka-python==2.0.2
prometheus-client==0.19.0
circuitbreaker==1.4.0
pydantic==2.5.0
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0
cryptography